% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAssign.R
\docType{package}
\name{Assignment}
\alias{Assignment}
\alias{Assignment-package}
\title{\code{Assignment} package
Please do not touch anything above the TEST_EX_X Lines!}
\description{
\code{Assignment} package
Please do not touch anything above the TEST_EX_X Lines!
}
