% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testAssign.R
\docType{package}
\name{testAssign}
\alias{testAssign}
\alias{testAssign-package}
\title{\code{testAssign} package
Please do not touch anything above the TEST_EX_X Lines!}
\description{
\code{testAssign} package
Please do not touch anything above the TEST_EX_X Lines!
}
